requirements.txt
streamlit==1.41.1
python-dotenv>=1.0
langchain-core>=0.2
langchain-community>=0.2
langchain-openai>=0.1
openai>=1.14
